requirements.txt
setuptools
wheel
django~=4.2.20
django-cors-headers
djangorestframework-simplejwt
djangorestframework-api-key @ git+https://github.com/LundIT/djangorestframework-api-key
django-filter
django-extensions
drf-oidc-auth
mozilla-django-oidc
XlsxWriter
scipy
numpy==1.26.4
pandas==1.4.3
openpyxl
psycopg2-binary
pdfkit
django-allow-cidr
django-cprofile-middleware
pretty_html_table
djangorestframework==3.15.2
requests
django-sendgrid-v5
redis
celery
webdriver-manager
sentry_sdk
channels
asyncio
channels-redis
uvicorn[standard]
nest_asyncio
Office365-REST-Python-Client
streamlit==1.44.0
streamlit-keycloak-lex
amqp
sqlalchemy
django_redis
django-storages[google]
pillow
DjangoSharepointStorage==1.1.3
click
django-lifecycle
python-keycloak==3.9.1
markdown
mistune
django-simple-history
markdown2
xhtml2pdf
tabulate
django-db-views
pre-commit
factory-boy
coverage
python-jose
